+++ lib/awesome_extensions.dart
@@ -7,11 +7,13 @@ export 'package:awesome_extensions/src/url_strategy.dart';
 export 'package:awesome_extensions/src/widgets/avatar_image.dart';
 import 'package:awesome_extensions/src/platform.dart';
 import 'package:awesome_extensions/src/shimmer.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 part 'context_extensions/media_query_extension.dart';
 part 'context_extensions/navigation_extension.dart';
 part 'context_extensions/theme_extension.dart';
+part 'context_extensions/dialog_extension.dart';
 part 'widget_extensions/shimmer_extension.dart';
 part 'widget_extensions/align_extension.dart';
 part 'widget_extensions/center_extension.dart';
@@ -24,3 +26,4 @@ part 'text_extensions/text_style_extension.dart';
 part 'double_extensions/size_box_extension.dart';
 part 'date_extensions/date_extension.dart';
 part 'num_extensions/number_extension.dart';
+part 'string_extensions/string_extension.dart';