+++ lib/string_extensions/string_extension.dart
@@ -1,0 +1,45 @@
+part of '../awesome_extensions.dart';
+
+extension StringExtension on String {
+  /// Checks if data is empty
+  bool isBlank() => this.isEmpty;
+
+  /// Capitalize each word inside string
+  /// Example: your name => Your Name, your name => Your name
+  String capitalize() {
+    if (this.isBlank()) return this;
+    return this.split(' ').map((e) => e.capitalize()).join(' ');
+  }
+
+  /// Uppercase first letter inside string and let the others lowercase
+  /// Example: your name => Your name
+  String capitalizeFirst() {
+    if (this.isBlank()) return this;
+    return this[0].toUpperCase() + this.substring(1).toLowerCase();
+  }
+
+  /// Remove all whitespace inside string
+  /// Example: your name => yourname
+  String removeAllWhitespace() {
+    return this.replaceAll(' ', '');
+  }
+
+  bool hasMatch(String pattern) {
+    return RegExp(pattern).hasMatch(this);
+  }
+
+  /// Checks if string consist only numeric.
+  /// Numeric only doesn't accepting "." which double data type have
+  bool isNumericOnly() => this.hasMatch(r'^\d+$');
+
+  /// Checks if string consist only Alphabet. (No Whitespace)
+  bool isAlphabetOnly() => this.hasMatch(r'^[a-zA-Z]+$');
+
+  /// Checks if string contains at least one Capital Letter
+  bool hasCapitalletter() => this.hasMatch(r'[A-Z]');
+
+  /// Checks if string is boolean.
+  bool isBool() {
+    return (this == 'true' || this == 'false');
+  }
+}